requirements.txt
pandas==1.1.3
streamlit==0.80.0
tabula_py==2.2.0
numpy
google-python-cloud-debugger